Include compilation errors in s:history for :Last
If the repl doesn't return a value due to an error during compilation
(EOF due to missing bracket, unrecognised symbol), the response has no
value and as a result is not added to the history. However it's useful
to add it so it can be inspected with :Last to aid with fixing the issue.