Add arglists to the omnicomplete preview window
When you use omnicomplete the arglists of functions are shown in the
popup menu and the docstrings in the preview window.  This patch adds
the arglists into the preview window too.  This is handy because once
you select a completion you'll still be able to see the arglists as you
write the rest of the form.