Fix sending interrupt commands after SIGINT
I was experimenting with this code and it appears that for some reason
the `catch Vim:Interrupt` statement was never executing. I couldn't
trace down why. This is sorta an ugly hack now, but I am not sure what
else to try catching if Vim:Interrupt isn't working.

It also appeared from my debugging that the `options` never had the
session in it, but `self` did, so I changed the code to copy the session
id from `self`.