Show exception causes in the location window
When evaluating an expression which throws an exception, the resulting
stacktrace frames are interpreted and assembled for display in the
location window: The root exception is shown along with all the causes.

The fallback mechanism (when the nREPL response doesn't supply a
stacktrace) has also been modified to produce the stacktrace causes
in exactly the same format.

Note that the 'old'-style nREPL response handling has not been
modified.