Fix run test under cursor in tests with metadata
The pattern used to find the test under cursor fails to match when the
deftest has metadata in the test var, like in
`(deftest ^:slow slow-test)`. Upon failure, it continues searching
backwards, usually matching a previous deftest or any other def.

The matching is fixed by adding an optional intermediate non-capturing
group that matches metadata added as `^:some-metadata`. Other ways to
add metadata will still make the pattern to fail, however they are not
so common when metadata is added to a test var.