Add support for piggieback 0.3.1+
The new piggieback can be found at https://github.com/clojure-emacs/piggieback

They have dropped support for Rhino, as it was supported by a number of
horrible hacks. Instead Nashorn is recommended as the default.

The old piggieback is still supported, and in the case you're using the old
piggieback & don't have nashorn available, you can manually call out to rhino.
The latest Clojure survey suggests that a large number of Clojure users do have
nashorn available.

Nashorn definitely works as of this commit, so close #243