Restore pretty print
The native nREPL support for pretty printing lacks an abstraction for
configuring options across different print functions, so I'm ripping out
the existing configuration options for now.  I'll entertain proposals
for a new configuration mechanic.

This is an attractive but nonessential feature, so defaulting to
fipp-pprint is a good way to bait users into installing Cider.  Checking
for the "info" op is an imperfect way to detect Cider, but in the event
that we're wrong, all that happens is an error message is included with
the response which we ignore.

Closes https://github.com/tpope/vim-fireplace/issues/329