Reduce visual "jumping" on successful RunTest
On my machine, the current implementation causes the quickfix window to
jump up and back down on a successful test run, and/or causes a "Press
ENTER to continue" prompt from the combination of the quickfix window
opening and echoing the test command.

This commit changes `copen` to `cwindow`, but moves the qflist clearing
to after that call. This keeps the quickfix window open if it was
already (there's a slight flash) on errors, but prevents the jumping on
a successful run. This also re-echos the test command after the results
come back to more clearly indicate a successful run, similar to how it
felt before the new async stuff.

I also tried conditionally using copen by checking the qflist contents
before clearing, but the list wasn't empty on a successful run, so there
might be something else hiding the window in that case that I'm not
aware of....