Show test progress after a short delay if tests are still running
Another stab at getting #347 right. With this commit, we will always
hide the quickfix window on success, and show it after 250ms *if* the
tests are still running. 50ms was too fast on my machine; it may make
sense to allow the delay to be customizable for slower machines.

With the "Started" message, the 250ms delay doesn't seem noticeable, but
that may be subjective. And, on a quick success, the delay doesn't even
come into play.