Fix catch 22 in nested nvim callbacks
Neovim refuses to process a second channel callback while a running
callback is sleeping.  This is inadequate when treating a channel as a
series of messages.  Hack around this by moving the meat of the callback
into a timer.

Neovim also won't run timers during a jobwait(), and the Vim solution of
sleep 1m causes weird redraw behavior, so we have to resort to wait()
instead.  Calling wait() directly with the max time was producing hard
to debug issues, so I've settled for calling it repeatedly with a 1
millisecond timeout.

Closes https://github.com/tpope/vim-fireplace/issues/378