Update 'cider-nrepl via CLJ' documentation link
Somewhat superflouous. The link to cider-nrepl's "With CLJ" documentation
appears to have changed, as it currently didn't point to the anchor tag it
was probably expecting. I've updated the link to now point to
https://docs.cider.mx/cider-nrepl/usage.html#via-clj which I'm largely
assuming is correct for this usecase and should be a more stable URL,
instead of the Readme in the GitHub repo.