transport#connect: pass nrepl+unix: strings unchanged
We need a way to let the user indicate that they want a AF_UNIX socket
connection. I don't have the Vimscript skills to plumb in an alternate
connection command and function, so here's a simple, dumb proof of
concept: if the user passes a string starting with nrepl+unix:, we'll
pass that string unmodified as the host value when we call the Python
script.