Enable Unix domain socket support
This patch allows users to connect to nREPL servers running on Unix
domain sockets by modifying :FireplaceConnect to let users provide
a connection string pointing to a socket file on the filesystem.

This is accomplished by modifying the setup code to identify such
connection strings and skip the URL building process and updating the
Python script to look for non-URL strings and attempt to connect to them
as socket type `AF_UNIX`.

For now, socket connections are assumed to be using the `nrepl` scheme,
and the current directory (or `b:java_root`) is used to populate the
"Scope connection to: " prompt.

Also removes the vestigial second argument to the Python script.